requirements.txt
python-telegram-bot>=21.0
supabase>=2.0
thefuzz[speedup]>=0.22
httpx>=0.27
python-dotenv>=1.0
